+++ _templates/new/entity/sidePanelImportDetailsPanel.ejs.t
@@ -3,4 +3,4 @@ inject: true
 to: src/containers/CrudEndpointUi/layout.js
 before: //hygen-inject-before4
 ---
-import <%= Name %>DetailsPanel from '../SidePanelDetails/<%= Name %>';
+import <%= className %>DetailsPanel from '../SidePanelDetails/<%= className %>';