+++ _templates/new/entity/sidePanelImportForm.ejs.t
@@ -3,4 +3,4 @@ inject: true
 to: src/containers/CrudEndpointUi/layout.js
 before: //hygen-inject-before3
 ---
-import <%= Name %>Form from '../Form/<%= Name %>';
+import <%= className %>Form from '../Form/<%= className %>';