+++ _templates/new/entity/sidePanelIndexTest.ejs.t
@@ -1,5 +1,5 @@
 ---
-to: src/containers/SidePanelDetails/<%= Name %>/__tests__/index.test.js
+to: src/containers/SidePanelDetails/<%= className %>/__tests__/index.test.js
 ---
 /*
  * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
@@ -9,7 +9,7 @@ import React from 'react';
 import { Map } from 'immutable';
 import { createStore } from 'redux';
 import { shallow } from 'enzyme';
-import <%= Name %>DetailsPanel, { mapStateToProps } from '../';
+import <%= className %>DetailsPanel, { mapStateToProps } from '../';
 import {
   getSelectedEntity,
   userHasUpdatePermission,
@@ -30,14 +30,14 @@ userHasUpdatePermission.mockImplementation(() => true);
 isInherited.mockImplementation(() => false);
 isSaving.mockImplementation(() => false);
 
-describe('<%= Name %>DetailsPanel Renders', () => {
+describe('<%= className %>DetailsPanel Renders', () => {
   it('renders', () => {
     const store = createStore(state => state);
     expect(
       shallow(
-        <<%= Name %>DetailsPanel store={store}>
+        <<%= className %>DetailsPanel store={store}>
           Child
-        </<%= Name %>DetailsPanel>
+        </<%= className %>DetailsPanel>
       )
     ).toMatchSnapshot();
   });